JDBC - Organize MySQL tests into MySQLDriverTest file
