+++ src/test/java/ortus/boxlang/runtime/bifs/global/jdbc/QueryExecuteTest.java
@@ -19,7 +19,6 @@
 package ortus.boxlang.runtime.bifs.global.jdbc;
 
 import static com.google.common.truth.Truth.assertThat;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -32,7 +31,6 @@ import java.util.List;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.EnabledIf;
 
 import ortus.boxlang.runtime.dynamic.casters.StructCaster;
 import ortus.boxlang.runtime.jdbc.ExecutedQuery;
@@ -48,81 +46,6 @@ import tools.JDBCTestUtils;
 public class QueryExecuteTest extends BaseJDBCTest {
 
 	static Key result = new Key( "result" );
-
-	@EnabledIf( "tools.JDBCTestUtils#hasMySQLModule" )
-	@DisplayName( "It supports timestamp param types" )
-	@Test
-	public void testTimestampDateParam() {
-		instance.executeSource(
-		    """
-		       queryExecute(
-		    	"
-		    	INSERT INTO developers ( id, name, role, createdAt )
-		    	VALUES ( 100, 'Tony Skipponi', 'Engineer', :timestamp )",
-		    	{
-		    		timestamp : { sqltype : "cf_sql_timestamp", value : now() }
-		    	},
-		    	{ "datasource" : "mysqldatasource" }
-		    );
-		    result = queryExecute( "SELECT * FROM developers WHERE id = 100", [], { "datasource" : "mysqldatasource" } );
-		       """,
-		    context );
-		assertThat( variables.get( result ) ).isInstanceOf( Query.class );
-		Query query = variables.getAsQuery( result );
-		assertEquals( 1, query.size() );
-	}
-
-	@EnabledIf( "tools.JDBCTestUtils#hasMySQLModule" )
-	@DisplayName( "It can use string values as timestamp params." )
-	@Test
-	public void testTimestampParamCompare() {
-		instance.executeSource(
-		    """
-		      queryExecute( "INSERT INTO developers ( id, name, role, createdAt )
-		      	VALUES ( 101, 'Tony Skipponi', 'Engineer', NOW() )", {}, { "datasource" : "mysqldatasource" } );
-		      result = queryExecute(
-		      	"
-		      	SELECT * FROM developers
-		    WHERE createdAt IS NOT NULL AND createdAt < :timestamp",
-		      	{
-		      		timestamp : { sqltype : "cf_sql_timestamp", value : "09/24/2099" }
-		      	},
-		      	{ "datasource" : "mysqldatasource" }
-		      );
-		      """,
-		    context );
-		assertThat( variables.get( result ) ).isInstanceOf( Query.class );
-		Query query = variables.getAsQuery( result );
-		assertEquals( 1, query.size() );
-	}
-
-	@EnabledIf( "tools.JDBCTestUtils#hasMySQLModule" )
-	@DisplayName( "It can execute multiple statements in a single queryExecute() call" )
-	@Test
-	public void testMultipleStatements() {
-		assertDoesNotThrow( () -> instance.executeStatement(
-		    """
-		           result = queryExecute( '
-		     	   TRUNCATE TABLE developers;
-		               INSERT INTO developers (id) VALUES (111);
-		               INSERT INTO developers (id) VALUES (222);
-		               SELECT * FROM developers;
-		               INSERT INTO developers (id) VALUES (333);
-		               INSERT INTO developers (id) VALUES (444);
-		               ',
-		      [],
-		      { "datasource" : "mysqldatasource" }
-		           );
-		    """, context )
-		);
-		Object multiStatementQueryReturn = variables.get( Key.of( "result" ) );
-		assertThat( multiStatementQueryReturn ).isInstanceOf( Query.class );
-		assertEquals( 2, ( ( Query ) multiStatementQueryReturn ).size(), "For compatibility, the last result should be returned" );
-
-		Query newTableRows = ( Query ) instance
-		    .executeStatement( "queryExecute( 'SELECT * FROM developers WHERE id IN (111,222)', [],{ 'datasource' : 'mysqldatasource' } );", context );
-		assertEquals( 2, newTableRows.size() );
-	}
 
 	@DisplayName( "It can execute a query with no bindings on the default datasource" )
 	@Test
